Tag dispatching to pxn_impl ctors
avr-gcc doesn't allow constexpr constructor with a body using
c++11. The usage of the constructos without the contsexpr label
increases the code size which is generated to a demo like led.cpp.

The solution was to use tag dispatching to the modes input, pullup and
output. This is something that is allowed to c++11 and in the end the
enum mode is replaced by tags like pullup_t that was already there.

+ Offers a method pxn_impl::is_low() to a better readability of the code.